pyproject.toml
[tool.poetry]
name = "fastapi-yamdb"
version = "0.1.0"
description = ""
authors = ["Daria"]
license = "BSD-3"
readme = "README.md"
packages = [{include = "fastapi_yamdb"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.83.0"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
SQLAlchemy = "^1.4.41"
psycopg2-binary = "^2.9.3"
alembic = "^1.8.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^1.10.2"}
python-multipart = "^0.0.5"
asyncpg = "^0.26.0"
databases = {extras = ["asyncpg"], version = "^0.6.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
